Major Changes
Created a file for settings. This file holds the colours used in the program.

Added to determine Patterns function. Now breaks up activity and allows iteration to compare and find patterns that go against user set rules.

Adding sleep to core file. This way its not spamming the command line with information.

Made it so when the user uses the monitor command it loops endlessly running every 30 seconds until they press Ctrl + C to stop monitoring the connection.